fixed issues

+++ PractiaAngular/src/app/customers-dashboard/dashboard/dashboard.component.ts
@@ -1,11 +1,15 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 
-import { MatTableDataSource } from '@angular/material/table';
+import { MatTableDataSource, MatTableModule } from '@angular/material/table';
+import { MatDialog } from '@angular/material/dialog';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatSort, MatSortModule } from '@angular/material/sort';
+
 import { Customers } from '../models/customers.interface';
 import { CustomersService } from '../services/customers.service';
-import { MatDialog } from '@angular/material/dialog';
 import { CustomerAddEditComponent } from '../customer-add-edit/customer-add-edit.component';
 import { CoreService } from '../core/core.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-dashboard',
@@ -25,6 +29,8 @@ export class DashboardComponent {
     'Phone',
     'Action',
   ];
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
+  @ViewChild(MatSort) sort!: MatSort;
 
   constructor(
     private _customerService: CustomersService,
@@ -36,29 +42,53 @@ export class DashboardComponent {
     this.getAllCustomers();
   }
 
+  ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
+    this.dataSource.sort = this.sort;
+    this.paginatorTranslation();
+  }
+
   getAllCustomers() {
     this._customerService.getAllCustomers().subscribe({
       next: (res) => {
         this.dataSource.data = res;
       },
       error: (err) => {
-        console.error();
+        this._coreService.openSnackBar(err.error.Message);
       },
     });
   }
 
   deleteCustomer(id: string) {
-    this._customerService.deleteCustomer(id).subscribe({
-      next: (res) => {
-        this._coreService.openSnackBar(
-          `Cliente con id "${id.trimEnd()}" eliminado`
-        );
-        this.getAllCustomers();
-      },
-      error: (err) => {
-        console.error();
-      },
-    });
+    this._dialog
+      .open(ConfirmDialogComponent, { width: '500px', data: id })
+      .afterClosed()
+      .subscribe({
+        next: (value) => {
+          if (value) {
+            this._customerService.deleteCustomer(id).subscribe({
+              next: (res) => {
+                this._coreService.openSnackBar(
+                  `Cliente con id "${id.trimEnd()}" eliminado`
+                );
+                this.getAllCustomers();
+              },
+              error: (err) => {
+                this._coreService.openSnackBar(err.error.Message);
+              },
+            });
+          }
+        },
+      });
+  }
+
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
+
+    if (this.dataSource.paginator) {
+      this.dataSource.paginator.firstPage();
+    }
   }
 
   openForm() {
@@ -85,5 +115,11 @@ export class DashboardComponent {
           }
         },
       });
+  }
+
+  paginatorTranslation() {
+    this.paginator._intl.itemsPerPageLabel = 'Elementos por pagina:';
+    this.paginator._intl.nextPageLabel = 'Siguiente pagina';
+    this.paginator._intl.previousPageLabel = 'Anterior pagina';
   }
 }
